requirements.txt
Django~=3.1.7
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0